refactor : Move Application to lotto package
