+++ src/test/java/lotto/ApplicationTest.java
@@ -1,4 +1,4 @@
-package domain;
+package lotto;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
